fix bigdata metadata to explicitly export functions required by config feeder client node info
