+++ granulate_utils/metadata/bigdata/interfaces.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class BigDataInfo:
+    provider: str
+    provider_version: str
